requirements.txt
flake8==7.1.1
mccabe==0.7.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
